cptbox: move CompilerIsolateTracer to own file
Also modify it to comply with new isolate.py changes.
