+++ dmoj/executors/compiled_executor.py
@@ -1,20 +1,15 @@
 import hashlib
 import os
 import pty
-import struct
-import sys
 from typing import Any, Dict, IO, List, Optional, Sequence
 
 import pylru
 
-from dmoj.cptbox import FilesystemSyscallKind, IsolateTracer, TracedPopen
-from dmoj.cptbox._cptbox import AT_FDCWD, Debugger
-from dmoj.cptbox.filesystem_policies import ExactFile, FilesystemAccessRule, RecursiveDir
-from dmoj.cptbox.handlers import ACCESS_EFAULT, ACCESS_EPERM, ALLOW
-from dmoj.cptbox.syscalls import *
-from dmoj.cptbox.tracer import AdvancedDebugger
+from dmoj.cptbox import TracedPopen
+from dmoj.cptbox.compiler_isolate import CompilerIsolateTracer
+from dmoj.cptbox.filesystem_policies import FilesystemAccessRule
 from dmoj.error import CompileError, OutputLimitExceeded
-from dmoj.executors.base_executor import BASE_FILESYSTEM, BASE_WRITE_FILESYSTEM, BaseExecutor, ExecutorMeta
+from dmoj.executors.base_executor import BaseExecutor, ExecutorMeta
 from dmoj.judgeenv import env
 from dmoj.utils.communicate import safe_communicate
 from dmoj.utils.error import print_protection_fault
@@ -71,164 +66,6 @@ class _CompiledExecutorMeta(ExecutorMeta):
             cls.compiled_binary_cache[cache_key] = obj
 
         return obj
-
-
-UTIME_OMIT = (1 << 30) - 2
-
-
-class CompilerIsolateTracer(IsolateTracer):
-    def __init__(self, tmpdir, read_fs, write_fs, *args, **kwargs):
-        read_fs += BASE_FILESYSTEM + [
-            RecursiveDir(tmpdir),
-            ExactFile('/bin/strip'),
-            RecursiveDir('/usr/x86_64-linux-gnu'),
-        ]
-        write_fs += BASE_WRITE_FILESYSTEM + [RecursiveDir(tmpdir)]
-        super().__init__(read_fs, *args, write_fs=write_fs, **kwargs)
-
-        self.update(
-            {
-                # Process spawning system calls
-                sys_fork: ALLOW,
-                sys_vfork: ALLOW,
-                sys_execve: ALLOW,
-                sys_getcpu: ALLOW,
-                sys_getpgid: ALLOW,
-                # Directory system calls
-                sys_mkdir: self.check_file_access('mkdir', 0, FilesystemSyscallKind.WRITE),
-                sys_mkdirat: self.check_file_access_at('mkdirat', FilesystemSyscallKind.WRITE),
-                sys_rmdir: self.check_file_access('rmdir', 0, FilesystemSyscallKind.WRITE),
-                # Linking system calls
-                sys_link: self.check_file_access('link', 1, FilesystemSyscallKind.WRITE),
-                sys_linkat: self.check_file_access_at('linkat', FilesystemSyscallKind.WRITE, file_reg=3),
-                sys_unlink: self.check_file_access('unlink', 0, FilesystemSyscallKind.WRITE),
-                sys_unlinkat: self.check_file_access_at('unlinkat', FilesystemSyscallKind.WRITE),
-                sys_symlink: self.check_file_access('symlink', 1, FilesystemSyscallKind.WRITE),
-                # Miscellaneous other filesystem system calls
-                sys_chdir: self.check_file_access('chdir', 0, FilesystemSyscallKind.READ),
-                sys_chmod: self.check_file_access('chmod', 0, FilesystemSyscallKind.WRITE),
-                sys_utimensat: self.do_utimensat,
-                sys_umask: ALLOW,
-                sys_flock: ALLOW,
-                sys_fsync: ALLOW,
-                sys_fadvise64: ALLOW,
-                sys_fchmodat: self.check_file_access_at('fchmodat', FilesystemSyscallKind.WRITE),
-                sys_fchmod: self.do_fchmod,
-                sys_fallocate: ALLOW,
-                sys_ftruncate: ALLOW,
-                sys_rename: self.do_rename,
-                sys_renameat: self.do_renameat,
-                # I/O system calls
-                sys_readv: ALLOW,
-                sys_pwrite64: ALLOW,
-                sys_sendfile: ALLOW,
-                # Event loop system calls
-                sys_epoll_create: ALLOW,
-                sys_epoll_create1: ALLOW,
-                sys_epoll_ctl: ALLOW,
-                sys_epoll_wait: ALLOW,
-                sys_epoll_pwait: ALLOW,
-                sys_timerfd_settime: ALLOW,
-                sys_eventfd2: ALLOW,
-                sys_waitid: ALLOW,
-                sys_wait4: ALLOW,
-                # Network system calls, we don't sandbox these
-                sys_socket: ALLOW,
-                sys_socketpair: ALLOW,
-                sys_connect: ALLOW,
-                sys_setsockopt: ALLOW,
-                sys_getsockname: ALLOW,
-                sys_sendmmsg: ALLOW,
-                sys_recvfrom: ALLOW,
-                sys_sendto: ALLOW,
-                # Miscellaneous other system calls
-                sys_msync: ALLOW,
-                sys_clock_nanosleep: ALLOW,
-                sys_memfd_create: ALLOW,
-                sys_rt_sigsuspend: ALLOW,
-            }
-        )
-
-        # FreeBSD-specific syscalls
-        if 'freebsd' in sys.platform:
-            self.update(
-                {
-                    sys_rfork: ALLOW,
-                    sys_procctl: ALLOW,
-                    sys_cap_rights_limit: ALLOW,
-                    sys_posix_fadvise: ALLOW,
-                    sys_posix_fallocate: ALLOW,
-                    sys_setrlimit: ALLOW,
-                    sys_cap_ioctls_limit: ALLOW,
-                    sys_cap_fcntls_limit: ALLOW,
-                    sys_cap_enter: ALLOW,
-                    sys_utimes: self.check_file_access('utimes', 0, FilesystemSyscallKind.WRITE),
-                }
-            )
-
-    def do_utimensat(self, debugger: AdvancedDebugger) -> bool:
-        timespec = struct.Struct({32: '=ii', 64: '=QQ'}[debugger.address_bits])
-
-        # Emulate https://github.com/torvalds/linux/blob/v5.14/fs/utimes.c#L152-L161
-        times_ptr = debugger.uarg2
-        if times_ptr:
-            try:
-                buffer = debugger.readbytes(times_ptr, timespec.size * 2)
-            except OSError:
-                return ACCESS_EFAULT(debugger)
-
-            times = list(timespec.iter_unpack(buffer))
-            if times[0][1] == UTIME_OMIT and times[1][1] == UTIME_OMIT:
-                debugger.syscall = -1
-
-                def on_return():
-                    debugger.result = 0
-
-                debugger.on_return(on_return)
-                return True
-
-        # Emulate https://github.com/torvalds/linux/blob/v5.14/fs/utimes.c#L142-L143
-        if debugger.uarg0 != AT_FDCWD and not debugger.uarg1:
-            path = self._getfd_pid(debugger.tid, debugger.uarg0)
-            return True if self.write_fs_jail.check(path) else ACCESS_EPERM(debugger)
-
-        return self.check_file_access_at('utimensat', FilesystemSyscallKind.WRITE)(debugger)
-
-    def do_fchmod(self, debugger: Debugger) -> bool:
-        path = self._getfd_pid(debugger.tid, debugger.uarg0)
-        return True if self.write_fs_jail.check(path) else ACCESS_EPERM(debugger)
-
-    def do_rename(self, debugger: Debugger) -> bool:
-        old_path, old_path_error = self.read_path('rename', debugger, debugger.uarg0)
-        if old_path_error is not None:
-            return old_path_error
-
-        new_path, new_path_error = self.read_path('rename', debugger, debugger.uarg1)
-        if new_path_error is not None:
-            return new_path_error
-
-        if not self._file_access_check(old_path, debugger, self.write_fs_jail):
-            return ACCESS_EPERM(debugger)
-        if not self._file_access_check(new_path, debugger, self.write_fs_jail):
-            return ACCESS_EPERM(debugger)
-
-        return True
-
-    def do_renameat(self, debugger: Debugger) -> bool:
-        old_path, old_path_error = self.read_path('renameat', debugger, debugger.uarg1)
-        if old_path_error is not None:
-            return old_path_error
-
-        new_path, new_path_error = self.read_path('renameat', debugger, debugger.uarg3)
-        if new_path_error is not None:
-            return new_path_error
-
-        if not self._file_access_check(old_path, debugger, self.write_fs_jail, dirfd=debugger.uarg0):
-            return ACCESS_EPERM(debugger)
-        if not self._file_access_check(new_path, debugger, self.write_fs_jail, dirfd=debugger.uarg2):
-            return ACCESS_EPERM(debugger)
-
-        return True
 
 
 class CompiledExecutor(BaseExecutor, metaclass=_CompiledExecutorMeta):
